Refactor: pull out StoreMenu
